+++ PetSpa.ModelViews/OrderModelViews/GetOrderViewModel.cs
@@ -13,7 +13,7 @@ namespace PetSpa.ModelViews.OrderModelViews
         public string? Id { get; set; }
         public string? Name { get; set; }
         public string? PaymentMethod { get; set; }
-        public double? Total { get; set; }
+        public decimal? Total { get; set; }
         public bool IsPaid { get; set; }
         public List<string>? OrderDetailId { get; set; }
         public string? CreatedBy { get; set; }
